rename packages and fix bug where records weren't emitted if cursor was not defined
