+++ airbyte_cdk/sources/declarative/requesters/query_properties/__init__.py
@@ -1,0 +1,14 @@
+# Copyright (c) 2025 Airbyte, Inc., all rights reserved.
+
+from airbyte_cdk.sources.declarative.requesters.query_properties.group_by_key import GroupByKey
+from airbyte_cdk.sources.declarative.requesters.query_properties.properties_from_endpoint import (
+    PropertiesFromEndpoint,
+)
+from airbyte_cdk.sources.declarative.requesters.query_properties.property_chunking import (
+    PropertyChunking,
+)
+from airbyte_cdk.sources.declarative.requesters.query_properties.query_properties import (
+    QueryProperties,
+)
+
+__all__ = ["GroupByKey", "PropertiesFromEndpoint", "PropertyChunking", "QueryProperties"]
