+++ airbyte_cdk/sources/declarative/requesters/query_properties/group_by_key.py
