+++ airbyte_cdk/sources/declarative/requesters/query_properties/properties_from_endpoint.py
