--- airbyte_cdk/sources/declarative/requesters/request_properties/__init__.py
@@ -1,14 +1,0 @@
-# Copyright (c) 2025 Airbyte, Inc., all rights reserved.
-
-from airbyte_cdk.sources.declarative.requesters.request_properties.group_by_key import GroupByKey
-from airbyte_cdk.sources.declarative.requesters.request_properties.properties_from_endpoint import (
-    PropertiesFromEndpoint,
-)
-from airbyte_cdk.sources.declarative.requesters.request_properties.property_chunking import (
-    PropertyChunking,
-)
-from airbyte_cdk.sources.declarative.requesters.request_properties.query_properties import (
-    QueryProperties,
-)
-
-__all__ = ["GroupByKey", "PropertiesFromEndpoint", "PropertyChunking", "QueryProperties"]
